[update] refactoring for context
